requirements.txt
asgiref==3.5.0
Django==3.2.12
gunicorn==20.1.0
django-ckeditor==6.4.0
django-js-asset==2.0.0
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0